+++ composer.json
@@ -20,13 +20,13 @@
         }
     ],
     "require": {
-        "php": "^8.3",
-        "filament/filament": "^4.0",
+        "php": "^8.4",
+        "filament/filament": "^5.0||^4.3.1",
         "owen-it/laravel-auditing": "^14.0.0",
-        "spatie/laravel-package-tools": "^1.15.0"
+        "spatie/laravel-package-tools": "^1.93.0"
     },
     "require-dev": {
-        "filament/upgrade": "^4.0",
+        "filament/upgrade": "^5.0",
         "laravel/pint": "^1.14",
         "nunomaduro/collision": "^8.1.1||^7.10.0",
         "larastan/larastan": "^3.0||^2.9",
@@ -52,6 +52,7 @@
     "scripts": {
         "post-autoload-dump": "@php ./vendor/bin/testbench package:discover --ansi",
         "analyse": "vendor/bin/phpstan analyse",
+        "upgrade": "vendor/bin/filament-v5",
         "test": "vendor/bin/pest",
         "test-coverage": "vendor/bin/pest --coverage",
         "format": "vendor/bin/pint"
