+++ src/Filament/Actions/Concerns/RestoresAudits.php
@@ -2,12 +2,12 @@
 
 namespace CrescentPurchasing\FilamentAuditing\Filament\Actions\Concerns;
 
-use Filament\Schemas\Components\Section;
-use Filament\Schemas\Components\Utilities\Get;
 use CrescentPurchasing\FilamentAuditing\Actions\GetAuditSchema;
 use CrescentPurchasing\FilamentAuditing\Actions\GetModifiedFields;
 use CrescentPurchasing\FilamentAuditing\Actions\RestoreAudit;
 use Filament\Forms\Components\Toggle;
+use Filament\Schemas\Components\Section;
+use Filament\Schemas\Components\Utilities\Get;
 use Livewire\Component;
 use OwenIt\Auditing\Models\Audit;
 
@@ -42,7 +42,11 @@ trait RestoresAudits
         $this->schema(function (Audit $record, GetAuditSchema $auditSchema, GetModifiedFields $fields): array {
             $keys = array_keys($record->new_values);
 
-            $fromSchema = $auditSchema($record->auditable?->toArray(), $keys);
+            if (! $record->auditable) {
+                return [];
+            }
+
+            $fromSchema = $auditSchema($record->auditable->toArray(), $keys);
 
             return [
                 Toggle::make('restore_to_old')
