--- src/Filament/Actions/Tables/ViewAuditableAction.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace CrescentPurchasing\FilamentAuditing\Filament\Actions\Tables;
-
-use Filament\Actions\Action;
-use CrescentPurchasing\FilamentAuditing\Filament\Actions\Concerns\ViewsAuditables;
-
-class ViewAuditableAction extends Action
-{
-    use ViewsAuditables;
-}
