--- src/Filament/Actions/Tables/ViewUserAction.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace CrescentPurchasing\FilamentAuditing\Filament\Actions\Tables;
-
-use Filament\Actions\Action;
-use CrescentPurchasing\FilamentAuditing\Filament\Actions\Concerns\ViewsUsers;
-
-class ViewUserAction extends Action
-{
-    use ViewsUsers;
-}
