+++ src/Filament/Actions/ViewAuditableAction.php
@@ -2,10 +2,29 @@
 
 namespace CrescentPurchasing\FilamentAuditing\Filament\Actions;
 
+use CrescentPurchasing\FilamentAuditing\Actions\GetAuditable;
 use CrescentPurchasing\FilamentAuditing\Filament\Actions\Concerns\ViewsAuditables;
 use Filament\Actions\Action;
+use OwenIt\Auditing\Models\Audit;
 
 class ViewAuditableAction extends Action
 {
-    use ViewsAuditables;
+    use ViewsAuditables {
+        setUp as baseSetUp;
+    }
+
+    protected function setUp(): void
+    {
+        $this->baseSetUp();
+
+        $this->iconButton();
+
+        $this->label(function (Audit $record, GetAuditable $getAuditable): string {
+            $title = $getAuditable->title($record) ?? '';
+
+            return __('filament-auditing::resource.actions.view.title', [
+                'title' => $title,
+            ]);
+        });
+    }
 }
