+++ src/Filament/Filters/QueryBuilder/AuditUserOperator.php
@@ -2,14 +2,14 @@
 
 namespace CrescentPurchasing\FilamentAuditing\Filament\Filters\QueryBuilder;
 
-use Filament\Schemas\Components\Grid;
-use Filament\Schemas\Components\Utilities\Set;
-use Filament\Schemas\Components\Utilities\Get;
 use CrescentPurchasing\FilamentAuditing\Actions\FormatAuditableType;
 use Exception;
 use Filament\Forms\Components\Field;
 use Filament\Forms\Components\MorphToSelect\Type;
 use Filament\Forms\Components\Select;
+use Filament\Schemas\Components\Grid;
+use Filament\Schemas\Components\Utilities\Get;
+use Filament\Schemas\Components\Utilities\Set;
 use Filament\Tables\Filters\QueryBuilder\Constraints\RelationshipConstraint\Operators\IsRelatedToOperator;
 use Illuminate\Config\Repository;
 use Illuminate\Database\Eloquent\Builder;
