+++ src/Filament/Resources/Audits/AuditResource.php
@@ -9,10 +9,10 @@ use CrescentPurchasing\FilamentAuditing\Actions\GetAuditSchema;
 use CrescentPurchasing\FilamentAuditing\Actions\GetModifiedFields;
 use CrescentPurchasing\FilamentAuditing\Actions\GetUser;
 use CrescentPurchasing\FilamentAuditing\Actions\GetUserSchema;
-use CrescentPurchasing\FilamentAuditing\Filament\Actions\Tables\RestoreAuditAction;
-use CrescentPurchasing\FilamentAuditing\Filament\Actions\Tables\ViewAuditableAction;
-use CrescentPurchasing\FilamentAuditing\Filament\Actions\Tables\ViewAuditAction;
-use CrescentPurchasing\FilamentAuditing\Filament\Actions\Tables\ViewUserAction;
+use CrescentPurchasing\FilamentAuditing\Filament\Actions\RestoreAuditAction;
+use CrescentPurchasing\FilamentAuditing\Filament\Actions\ViewAuditableAction;
+use CrescentPurchasing\FilamentAuditing\Filament\Actions\ViewAuditAction;
+use CrescentPurchasing\FilamentAuditing\Filament\Actions\ViewUserAction;
 use CrescentPurchasing\FilamentAuditing\Filament\Filters\QueryBuilder\AuditUserConstraint;
 use CrescentPurchasing\FilamentAuditing\Filament\Filters\QueryBuilder\AuditUserOperator;
 use CrescentPurchasing\FilamentAuditing\Filament\RelationManagers\AuditsRelationManager;
@@ -25,7 +25,6 @@ use Filament\Actions\ActionGroup;
 use Filament\Forms\Components\DateTimePicker;
 use Filament\Forms\Components\Textarea;
 use Filament\Forms\Components\TextInput;
-use Filament\Forms\Contracts\HasForms;
 use Filament\Resources\Resource as FilamentResource;
 use Filament\Schemas\Components\Grid;
 use Filament\Schemas\Components\Tabs;
