+++ src/Filament/Resources/Audits/Schemas/AuditForm.php
@@ -2,7 +2,4 @@
 
 namespace CrescentPurchasing\FilamentAuditing\Filament\Resources\Audits\Schemas;
 
-class AuditForm
-{
-
-}
+class AuditForm {}
