+++ tests/src/Filament/Actions/ViewAuditableActionTest.php
@@ -1,10 +1,11 @@
 <?php
 
-use CrescentPurchasing\FilamentAuditing\Filament\Actions\Tables\ViewAuditableAction;
+use CrescentPurchasing\FilamentAuditing\Filament\Actions\ViewAuditableAction;
 use CrescentPurchasing\FilamentAuditing\Filament\Resources\Audits\Pages\ManageAudits;
 use CrescentPurchasing\FilamentAuditing\Tests\Filament\Resources\Article\ArticleResource;
 use CrescentPurchasing\FilamentAuditing\Tests\Models\Article;
 use OwenIt\Auditing\Models\Audit;
+
 use function Pest\Livewire\livewire;
 
 it('Can see the url of the audit record', function () {
