+++ tests/src/Filament/Actions/ViewUserActionTest.php
@@ -1,10 +1,11 @@
 <?php
 
-use CrescentPurchasing\FilamentAuditing\Filament\Actions\Tables\ViewUserAction;
+use CrescentPurchasing\FilamentAuditing\Filament\Actions\ViewUserAction;
 use CrescentPurchasing\FilamentAuditing\Filament\Resources\Audits\Pages\ManageAudits;
 use CrescentPurchasing\FilamentAuditing\Tests\Filament\Resources\User\UserResource;
 use CrescentPurchasing\FilamentAuditing\Tests\Models\Article;
 use OwenIt\Auditing\Models\Audit;
+
 use function Pest\Livewire\livewire;
 
 it('Can see the url of the audit record', function () {
